==> explore/1_collect_traits_fishbase.Rmd <==
---
title: "Collect traits data from FishBase and SeaLifeBase"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)

library(oharac)
oharac::setup()
# remotes::install_github("ropensci/rfishbase")
library(rfishbase)
# Sys.setenv(FISHBASE_API="fishbase")
library(ggfortify) # For PCA biplot

```

# Summary

Get trait data for a select set of traits from FishBase and SeaLifeBase using the `rfishbase` package.

# Data

FishBase is a scientific database, and this has the implication - among others - that its use and the use of its contents are free as long as due credit is given.

This may be done at different levels, for which we suggest different forms of citations:

* when referring to FishBase concepts and design, cite its architects (Froese and Pauly 2000);
* when referring to a set of values extracted from a FishBase table, cite the author(s) of the original data, e.g., "Houde and Zastrow (1993)", or "Welcomme (1988)". To help us track the use of FishBase in the literature, we would appreciate your also citing Froese and Pauly (2000) in an appropriate part of your text, as the source of the information;
* when discussing the features of a FishBase table, cite the section documenting that table, e.g., "Sa-a et al. (2000)."

## References

* Froese, R. and D. Pauly, Editors. 2000. FishBase 2000: concepts, design and data sources. ICLARM, Los Baños, Laguna, Philippines. 344 p.
* Houde, E.D. and C.E. Zastrow. 1993. Ecosystem- and taxon-specific dynamic energetics properties of fish larvae assemblages. Bull. Mar. Sci. 53(2):290-335.
* Sa-a, P., M.L. Palomares and D. Pauly. 2000. The FOOD ITEMS table, p. 182-188. In R. Froese and D. Pauly (eds.) FishBase 2000: concepts, design and data sources. ICLARM, Los Baños, Laguna, Philippines. 344 p.
* Welcomme, R.L. 1988. International introductions of inland aquatic species. FAO Fish. Tech. Pap. 294, 318 p.
    
# Methods

## Read in `species()` data for both fishbase and sealifebase

```{r}
get_fb_slb <- function(fun, 
                       keep_cols = NULL, keep_fxn = contains, 
                       drop_cols = NULL, drop_fxn = all_of) {
  fb <- fun(server = 'fishbase') %>%
    janitor::clean_names() %>%
    mutate(db = 'fb')
  slb <- fun(server = 'sealifebase') %>%
    janitor::clean_names() %>%
    mutate(db = 'slb')
  if(!is.null(keep_cols)) {
    fb <- fb %>%
      select(spec_code, species, db, keep_fxn(keep_cols))
    slb <- slb %>%
      select(spec_code, species, db, keep_fxn(keep_cols))
  }
  if(!is.null(drop_cols)) {
    fb <- fb %>%
      select(-drop_fxn(drop_cols))
    slb <- slb %>%
      select(-drop_fxn(drop_cols))
  }
  ### sometimes class of a column from one server mismatches that of the other.
  ### This should only be a problem for character vs. numeric - i.e.,
  ### not a problem for numeric vs. integer
  cols_check <- data.frame(
    type_fb = sapply(fb, class),
    col = names(fb)) %>%
    inner_join(data.frame(
      type_slb = sapply(slb, class),
      col = names(slb))) %>%
    filter(type_fb != type_slb) %>%
    filter(type_fb == 'character' | type_slb == 'character')
  if(nrow(cols_check) > 0) {
    message('Conflicting column types; coercing all to character:')
    message(cols_check)
    fb <- fb %>%
      mutate(across(all_of(cols_check$col), as.character))
    slb <- slb %>%
      mutate(across(all_of(cols_check$col), as.character))
  }

  return(bind_rows(fb, slb))
}
```

```{r}

keep_cols <- c('f_bname', 'length', 'depth', 'demers', 'pd50', 'longev')
species_df <- get_fb_slb(species, 
                         keep_cols = keep_cols, keep_fxn = contains,
                         drop_cols = '_ref',    drop_fxn = ends_with)

```

### instances of data for each column

`r sapply(species_df, FUN = function(x) sum(!is.na(x)))`

## Get length data

### examine columns and identify gapfill potential

Many species (~28%) have records for `length`, while a handful (~11%) have records for `common_length`.  But there are 312 spp (~.2%) with `common_length` and no `length`, and 709 with `length_female` but no `length`.

Can we gapfill some of these?  Note: Leatherbacks have 916000 cm as `common_length`, so eliminate that as an obvious error.
```{r compare length - common_length - length_female}

species_df %>%
  filter(is.na(length)) %>%
  summarize(com_no_f = sum(is.na(length_female) & !is.na(common_length)),
            f_no_com = sum(!is.na(length_female) & is.na(common_length)),
            both = sum(!is.na(length_female) & !is.na(common_length)))
species_df_outlier_removed <- species_df %>%
  filter(common_length < 2500)
ggplot(species_df_outlier_removed, aes(x = length, y = common_length)) +
  geom_point() +
  geom_point(aes(y = length_female), color = 'pink')

x <- lm(common_length ~ length, data = species_df_outlier_removed)

y <- lm(length_female ~ length, data = species_df)

l_c_coeff <- x$coefficients[2]
l_f_coeff <- y$coefficients[2]
```
`common_length ~ length` shows intercept of less than 2 cm, coefficient of .54, with p < 2e-16 and R^2^ of 86%.  This has strong potential to gapfill max length to gain an additional ~300 spp.

`length_female ~ length` shows an intercept of 3 cm, a coefficient of .89, with highly significant p values and R^2^ around 92%.  This has strong potential to gapfill max length ~700 spp.

Only 11 spp have no `length` but both `common_length` and `length_female`.

### gather length information

```{r}

check_df <- species_df %>%
  filter(!is.na(length_female) & !is.na(common_length)) %>%
  mutate(check_f = length_female / l_f_coeff,
         check_c = common_length / l_c_coeff)

length_trait <- species_df %>%
  mutate(l_value = case_when(!is.na(length) ~ length,
                             !is.na(length_female) ~ length_female / l_f_coeff,
                             !is.na(common_length) ~ common_length / l_c_coeff,
                             TRUE ~ NA_real_)) %>%
  mutate(l_column = case_when(!is.na(length) ~ 'length',
                              !is.na(length_female) ~ 'length_female',
                              !is.na(common_length) ~ 'common_length',
                              TRUE ~ NA_character_)) %>%
  select(db, spec_code, species, l_value, l_column) %>%
  distinct()

```

## Position in water column

score 0 = benthic, 1 = demersal, 2 = pelagic - from bottom up
```{r}
water_col_df <- species_df %>%
  select(db, spec_code, species, demers_pelag) %>%
  mutate(demers_pelag = tolower(demers_pelag))

table(water_col_df$demers_pelag)

water_col_trait <- water_col_df %>%
  mutate(water_col = case_when(str_detect(demers_pelag, 'benthic') ~ 1,
                               str_detect(demers_pelag, 'sessile') ~ 1,
                               str_detect(demers_pelag, 'demersal|benthopelagic') ~ 2,
                               str_detect(demers_pelag, 'reef') ~ 2, ### assume reef spp are demersal
                               str_detect(demers_pelag, 'pelagic') ~ 3,
                               TRUE ~ NA_real_))
```

## Phylogenetic diversity

Probably not much discerning power here - most are around 0.50.  

```{r, eval = FALSE}
pd50_df <- species_df %>%
  select(db, spec_code, species, pd50)

hist(pd50_df$pd50)
summary(pd50_df$pd50)
```

## Longevity

Too many NAs - not enough spp here to bother with...

```{r, eval = FALSE}
longev_df <- species_df %>%
  select(db, spec_code, species, longevity_wild, longevity_captive)

summary(longev_df$longevity_wild)
summary(longev_df$longevity_captive)
```

## Food:

```{r food table, eval = FALSE}
food_df <- get_fb_slb(fun = fooditems, keep_cols = c('food_i', 'troph'), keep_fxn = contains) 
```

### food_i

`r table(food_df$food_i)`

### food_ii

`r table(food_df$food_ii)`

### food_iii

`r table(food_df$food_iii)`

## Diet

Trophic level only for 2480 fish and sealife spp - mostly higher trophic.  Use this combined with the `food` columns above?

``` {r diet table, eval = FALSE}
diet_df <- get_fb_slb(fun = diet, keep_cols = 'troph', keep_fxn = contains)

hist(diet_df$troph)
```

```{r, eval = FALSE}
x <- diet_df %>%
  group_by(db, spec_code, species) %>%
  summarize(troph = mean(troph, na.rm = TRUE))
y <- x %>%
  left_join(food_df)
```

## Ecology table:
``` {r trophic ecology table}
fields <- c('troph', 'diet', 'school', 'shoal', 'sessile', 
            'oceanic', 'pelagic', 'demersal', 'benthic',
            'neritic', 'littoral', 'reef')

ecol <- get_fb_slb(fxn = ecology,
                   keep_cols = fields, keep_fxn = contains,
                   drop_cols = '_ref', drop_fxn = ends_with)
```

`food_troph` seems to have the most records of the troph columns, ~10000.
```{r}
ecol_troph_df <- ecol %>%
  select(db, spec_code, species, contains('troph')) # %>%
  # left_join(diet_df %>% group_by(species, spec_code, db) %>% summarize(troph = mean(troph, na.rm = TRUE)))
summary(ecol_troph_df)

ecol_troph_trait <- ecol_troph_df %>%
  select(db, spec_code, species, diet_troph, food_troph)
```

```{r}
gregar_df <- ecol %>%
  select(spec_code, species, db, contains(c('shoal', 'school')))
summary(gregar_df)
table(gregar_df$shoaling)
table(gregar_df$schooling)

gregar_trait <- gregar_df %>%
  select(spec_code, species, db, shoaling, schooling) %>%
  mutate(shoaling = abs(shoaling), schooling = abs(schooling))
```


## [Reproduction](https://www.fishbase.se/manual/English/fishbasereproduction00002700.htm)

Fish display an astonishing variety of reproductive modes, ranging from parthenogenesis in the molly Poecilia formosa to permanently attached parasitic males in the deep-sea fish Haplophryne mollis. Similarly, fecundity ranges from 300 million eggs per year in Mola mola to a few live born offspring, e.g., in many sharks (Lagler et al. 1977). Parental care may be absent, as in many pelagic fishes, or involve various kinds of nest guarding or mouthbrooding. This variety implies that constraints to the ability of fish populations to reproduce themselves will also take different forms. Knowledge of reproduction is therefore important for proper management and conservation of fish species.

Information on reproduction in FishBase is assembled in three tables: REPRODUCTION, MATURITY and SPAWNING. 

* [The REPRODUCTION table](https://www.fishbase.se/manual/English/fishbasethe_reproduction_table.htm) documents the general mode and the type of reproduction that apply to the species throughout its range. 
* [The MATURITY](https://www.fishbase.se/manual/English/fishbasethe_maturity_table.htm) and [SPAWNING tables](https://www.fishbase.se/manual/English/fishbasethe_spawning_table.htm), on the other hand, present information on the size and age at first maturity and spawning of different populations of the same species, occurring at different localities. These tables are described below.)

### reproduction

```{r}
fields <- c('fertilization', 'parental', 'guild', 'repro_mode', 'spawn')

reprod <- get_fb_slb(fun = reproduction,
                     keep_cols = fields, keep_fxn = contains,
                     drop_cols = '_ref', drop_fxn = ends_with)
summary(reprod)

# table(reprod$parental_care)
# biparental   maternal       none   paternal 
#         71        296       5163       1231 
# table(reprod$parental_care_q)
#    1    2    3    4 
# 3909  135 2589  126 
# table(reprod$fertilization)
 #           external in brood pouch or similar structure                            in mouth 
 #              25856                                 367                                  86 
 # internal (oviduct)                               other 
 #               8603                                   6 
# table(reprod$rep_guild1)
    # bearers    guarders nonguarders 
    #   15998        2711        9918 
# table(reprod$rep_guild2)
#                         brood hiders                         Brood hiders 
#                                  295                                    8 
#                       clutch tenders                       Clutch tenders 
#                                  374                                    5 
#                    external brooders                    External brooders 
#                                 1054                                13821 
#                internal live bearers                Internal live bearers 
#                                  961                                   97 
#                              nesters                              Nesters 
#                                 1577                                  394 
#       Open substratum egg scatterers open water/substratum egg scatterers 
#                                 2557                                 3925 
# Open water/substratum egg scatterers 
#                                 1126 
# table(reprod$repro_mode)
#       binary fission             dioecism      parthenogenesis            protandry            protogyny 
#                    2                38397                    3                 1425                  505 
# true hermaphroditism 
#                   65 
```
Traits for functional group could be:

* `parental_care`, coded as 0 (none), 1 (maternal or paternal), 2 (biparental): 
```
biparental   maternal       none   paternal 
        71        296       5163       1231 
```
* `rep_guild1`, coded as 0 (non-guarders), 1 (guarders), 2 (bearers)
```
bearers    guarders nonguarders 
  15998        2711        9918 
```

### maturity

So many NAs in this table.

``` {r maturity table}

fields <- c('age', 'length', 'tm', 'lm')

maturity <- get_fb_slb(fun = maturity,
                     keep_cols = fields, keep_fxn = contains,
                     drop_cols = '_ref', drop_fxn = ends_with)
summary(maturity)


```

### spawning

So many NAs here as well.

```{r spawning table}
fields <- c('fecun', 'length', 'weight')

spawn <- get_fb_slb(fun = spawning,
                     keep_cols = fields, keep_fxn = contains,
                     drop_cols = '_ref', drop_fxn = ends_with)
summary(spawn)
```

### Also in here: 

### Fecundity

duplicated from spawning/reproduction tables?  Again, so many NAs

```{r fecundity table}
fields <- c('fecun', 'length', 'weight')
fecund <- get_fb_slb(fun = fecundity,
                     keep_cols = fields, keep_fxn = contains,
                     drop_cols = '_ref', drop_fxn = ends_with)
summary(fecund)
```

### Ecosystem

doesn't look like traits so much as ecosystem types, incl MEOW.  
``` {r ecosystem table, eval = FALSE}
ecosys <- ecosystem()
### 155811 obs of 99 vars
### looks like MEOW regions and the like, E_CODE
DT::datatable(summarize_table(ecosys))

```

## [Population Dynamics](https://www.fishbase.se/manual/English/fishbasepopulation_dynamics00002680.htm)

Information on the maximum size and age of fish, on their length-weight relationships and estimates of their growth parameters, natural mortality and recruitment variability are crucial for fisheries management purposes.

While maximum age and size, and length-weight relationships are relatively easy to obtain for most fish species, making sure that such information is available wherever and whenever needed - and in the appropriate format - is rather more difficult.

* [The POPCHAR Table](https://www.fishbase.se/manual/English/fishbasethe_popchar_table.htm) 
    * This table presents information on maximum length (Lmax), weight (Wmax) and age (tmax) from various localities where a species occurs. The largest values from this table are also entered in the SPECIES table. The POPCHAR table also indicates whether the Lmax, Wmax and tmax values or various combinations thereof refer to the same individual fish.
    * `popchar()` - local population values, rather than global
* [The LENGTH-WEIGHT Table](https://www.fishbase.se/manual/English/fishbasethe_length_weight_table.htm)
    * Length-weight relationships are important in fisheries science, notably to raise length-frequency samples to total catch, or to estimate biomass from underwater length observations. The LENGTH-WEIGHT table presents the a and b values of over 5,000 length-weight relationships of the form W = a × Lb, pertaining to about over 2,000 fish species.
    * The units of length and weight in FishBase are centimeter and gram, respectively. Thus when length-weight relationships are not in cm-g, the intercept 'a' is transformed as follows...
    * `length_weight()`
    * potentially use for a proxy for aspect ratio?
* [The LENGTH-FREQUENCY Table](https://www.fishbase.se/manual/English/lengthfrequency.htm)
    * Length-frequency data are widely used to derive growth estimates, especially in small tropical fishes (see the POPGROWTH table, this vol.). Froese and Binohlan (2000) have shown that length-frequency curves can also be used to get a first assessment of the status of a stock (see Fig. 33) if the data are plotted in a framework of asymptotic length, length at optimum yield, and length at first maturity (see Key Facts, this vol.). With this new table, we try to collect and preserve historical data from unfished or still lightly fished populations, to be contrasted with the curves typically produced from overexploited stocks, where the large, highly fecund fish (the ‘Mega-spawners’) have disappeared and the bulk of the catch is made up of juveniles which had no chance to reproduce.
    * `length_freq()`
* [The LENGTH-LENGTH Table](https://www.fishbase.se/manual/english/PDF/FB_Book_CBinohlan_Length-Length_RF_JG.pdf)
    * This table contains relationships for the conversion of one length type to another for over 8,000 species of fish, derived from different publications, e.g. Moutopoulos and Stergiou (2002) and Gaygusuz et al (2006), or from fish pictures, e.g. Collette and Nauen (1983), Compagno (1984) and Randall (1997)
    * `length_length()`
* [The POPGROWTH Table](https://www.fishbase.se/manual/English/fishbasethe_popgrowth_table.htm)
    * This table contains information on growth, natural mortality and length at first maturity, which serve as inputs to many fish stock assessment models. The data can also be used to generate empirical relationships between growth parameters or natural mortality estimates, and their correlates (e.g., body shape, temperature, etc.), a line of research that is useful both for stock assessment and for increasing understanding of the evolution of life-history strategies (see Fig. 19).
    * `popgrowth()`
* [The RECRUITMENT Table](https://www.fishbase.se/manual/English/fishbasethe_recruitment_table.htm)
    * Recruitment fluctuations and their causes represent one of the main areas of research in fisheries science, which is not surprising since it is these fluctuations which determine the annual catch levels of fisheries.
    * Precise prediction of future recruitment is not possible. However, broad generalizations are possible (e.g., that depleted stocks produce fewer recruits than healthy stocks, a non-trivial result). The more recruitment time series are available from various parts of the world, the more precise and reliable will the generalizations be.
    * no function?

### PopChar
skip it for now

### Length-Weight

Explore for use as morphology continuous variable a la aspect ratio... units cm and g; formula:

$$W = aL^b$$

Can we compare body shapes to $a$ and $b$ parameters, to see how particular shapes correspond to these parameters (and then use an aspect ratio in place of descriptive shapes)

``` {r length-weight table, eval = FALSE}
lw <- length_weight()
### 46082 obs of 38 vars
DT::datatable(summarize_table(lw))

fields <- c('^a$', '^b$', 'length')

lw_select <- clean_df(lw, fields)

```

### Length-Length
skip for now

### Pop Growth

"This table contains information on growth, natural mortality and length at first maturity..." params for VBGF:

$$L_t = L_{\infty} \left(1 - e^{-K(t - t_0)}\right)$$

``` {r pop growth table, eval = FALSE}
fields <- c('loo', 'lm', 'length', 'to', '^k$')

popg <- get_fb_slb(fun = popgrowth,
                     keep_cols = fields, keep_fxn = contains,
                     drop_cols = '_ref', drop_fxn = ends_with)
summary(popg)

```

## Combine 'em

For a test run, let's try length, 
```{r}
traits_df <- length_trait %>%
  full_join(ecol_troph_trait) %>%
  full_join(gregar_trait) %>%
  full_join(water_col_trait) %>%
  mutate(log_l = log(l_value),
         gregar = schooling | shoaling,
         troph  = ifelse(is.na(diet_troph), food_troph, diet_troph),
         across(where(is.logical), as.numeric)) %>%
  ### drop cols that have been aggregated
  select(-food_troph, -diet_troph, -schooling, -shoaling, -l_value) %>%
  # group_by(db, spec_code, species) %>%
  # summarize(across(where(is.numeric), .fns = ~mean(.x, na.rm = TRUE))) %>%
  # ungroup() %>%
  distinct() %>%
  drop_na()

# ggplot(traits_df, aes(x = log_l, y = water_col)) +
#   geom_point()
```

``` {r PCA}
traits_pca <- traits_df %>%
  select(-where(is.character), -spec_code) %>%
  scale() %>%
  prcomp()

autoplot(traits_pca,
         data = traits_df,
         loadings = TRUE,
         colour = 'db',
         loadings.label = TRUE,
         loadings.colour = "black",
         loadings.label.colour = "black",
         loadings.label.vjust = -0.5
         ) +
  scale_color_manual(values = c("blue","purple","orange")) +
  scale_fill_manual(values = c("blue","purple","orange")) +
  theme_minimal()

# Variance explained by each PC
screeplot(traits_pca, type = "lines")

# See the loadings (weighting for each principal component)
traits_pca$rotation %>% round(3)

```

```{r kmeans}

traits_kmeans <- traits_df %>%
  select(-where(is.character), -spec_code) %>%
  scale() %>%
  as.matrix() %>%
  kmeans(8)

autoplot(traits_kmeans,
         data = traits_df,
         loadings = TRUE,
         colour = 'db',
         loadings.label = TRUE,
         loadings.colour = "black",
         loadings.label.colour = "black",
         loadings.label.vjust = -0.5
         ) +
  scale_color_manual(values = c("blue","purple","orange")) +
  scale_fill_manual(values = c("blue","purple","orange")) +
  theme_minimal()

traits_df_kmeans <- traits_df %>%
  mutate(cluster = traits_kmeans$cluster)

traits_sum <- traits_df_kmeans %>%
  group_by(cluster) %>%
  select(-spec_code) %>%
  summarize(across(is.numeric, .fns = list(mean = mean, sd = sd)), 
            n_tot = n(), n_fb = sum(db == 'fb'), n_slb = sum(db == 'slb'))
```
