==> explore/1_create_spp_str_intsx.Rmd <==
---
title: "Create vuln-weighted species-stressor intersections"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')

```

# Summary

Get species ranges; match to spp vulnerability scores.  

* For each stressor type, make a map of all spp ranges, weighted by the spp vulnerability, to calculate the mean predicted vulnerability to stressors in each pixel (and other distribution parameters).  This is vulnerability only - not impact.
* For each stressor intensity layer, for each spp, calculate a map of vulnerability-weighted impacts.
    * For multiple stressors, calculate cumulative impact to species pixel by pixel.  
    * Map mean cumulative impact across all spp.
    * For high- and low-cumulative-impact locations, make a plot showing stacked bars for each spp present, with what stressors are particularly impactful.
    
# Methods

## Set up file locations

```{r}
dir_spp_maps  <- '~/git-annex/bd_chi/spp_rasts_mol_2020'
dir_str_rasts <- '~/git-annex/bd_chi/layers/stressors_10km'
dir_intsx     <- '~/git-annex/spp_vuln/spp_str_intsx'
```

## Identify matched spp in IUCN

```{r}
### note a lot of these columns start with a bunch of NAs; use 
vuln_all_gapfilled <- data.table::fread('~/git-annex/spp_vuln/vuln_all_spp_gapfilled.csv') %>%
  select(-c(kingdom:genus))

iucn_spp <- read_csv('_data/spp_marine_maps_2020-1.csv') %>%
  mutate_if(is.character, tolower)

vuln_iucn <- iucn_spp %>%
  inner_join(vuln_all_gapfilled, by = c('sciname' = 'species')) %>%
  filter(presence != 5) %>%
  select(iucn_sid, sciname, taxon, stressor, vuln, match_rank, gapfill) %>%
  distinct()

vuln_iucn$iucn_sid %>% n_distinct() ### 6173 easy matches
```

## Loop over species and stressors

For each species/stressor combo:

* pull spp map
* pull str map
* calculate spp_str_intsx as presence(spp) * vuln(spp-str) * intensity(str)
* write out as spp_str_intsx map

```{r helper fxns}
cell_id_rast <- raster('_spatial/cell_id_mol.tif')
get_spp_range <- function(id) {
  message('getting species range raster for ', id)
  range_f <- file.path(dir_spp_maps, sprintf('iucn_sid_%s.csv', id))
  range_df <- read_csv(range_f, col_types = 'ii') %>%
    filter(presence != 5) %>%
    mutate(presence = 1) ### coerce all remaining presence to 1
  range_rast <- subs(cell_id_rast, range_df, by = 'cell_id', which = 'presence')
  return(range_rast)
}

get_str_rast <- function(s) {
  ### for a given intersection file, 
  str_chi <- read_csv('_raw/strs_vuln_to_chi.csv', col_types = 'cc') %>%
    filter(str_vuln == s) %>%
    .$str_chi %>%
    str_split(';') %>%
    unlist()
  str_rast_fs <- sprintf('%s_2013_rescaled_mol10km.tif', str_chi)
  str_stack <- raster::stack(file.path(dir_str_rasts, str_rast_fs))

  if(nlayers(str_stack) > 1) {
    str_rast <- calc(str_stack, sum, na.rm = TRUE)
    values(str_rast)[values(str_rast) == 0] <- NA
  } else {
    str_rast <- str_stack[[1]]
  }

  return(str_rast)
}

get_spp_str_vuln <- function(id, s) {
  v <- vuln_iucn %>%
    filter(iucn_sid == id & stressor == s) %>%
    .$vuln
  if(is.na(v)) stop('NA vulnerability for ', s, ' for ', id)
  return(v)
}

rast_to_df <- function(rast, name = 'value') {
  df <- data.frame(cell_id = 1:ncell(rast),
                   value = values(rast)) %>%
    filter(!is.na(value)) %>%
    rename(!!name := value)
  if(nrow(df) == 0) {
    df <- data.frame(cell_id = NA) %>%
      mutate(value = NA) %>%
      rename(!!name := value)
  }
  return(df)
}
```

```{r}
spp_ids <- vuln_iucn$iucn_sid %>% unique() %>% sort()
strs_vuln_to_chi <- read_csv('_raw/strs_vuln_to_chi.csv', col_types = 'cc') %>%
  filter(!is.na(str_chi))
strs_vuln <- strs_vuln_to_chi$str_vuln
for(id in spp_ids) {
  # id <- 155123
  message('Processing species ID ', id)
  intsx_fname <- sprintf('spp_str_intsx_%s_%s.csv', id, strs_vuln)
  intsx_files <- file.path(dir_intsx, intsx_fname)
  
  if(all(file.exists(intsx_files))) {
    ### skip to the end
    next()
  }
  
  spp_range_rast <- get_spp_range(id)
  
  x <- parallel::mclapply(intsx_files, mc.cores = 9,
                          FUN = function(f) {
  # for(f in intsx_files) { # f <- intsx_files[1]
    if(file.exists(f)) {
      message('file exists: ', f, '... skipping!')
    } else {
      message('Calculating ', f)
      s <- str_replace_all(basename(f), '.+[0-9]+_|.csv', '')
      str_rast <- get_str_rast(s)
      spp_str_vuln <- get_spp_str_vuln(id, s)
      if(spp_str_vuln > 0) {
        spp_range_vuln <- spp_range_rast * spp_str_vuln 
        spp_str_impact <- spp_range_vuln * str_rast
      } else {
        spp_str_impact <- raster(str_rast) ### NA raster
      }
      spp_str_impact_df <- rast_to_df(spp_str_impact, name = 'impact') %>%
        mutate(impact = round(impact, 5))
      write_csv(spp_str_impact_df, f)
    }
  })
}
```
